% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare}
\alias{compare}
\title{Compare the ungrouped with the grouped results}
\usage{
compare(df)
}
\arguments{
\item{df}{(dataframe dash or wages)}
}
\value{
A scatterplot, some intepretations and an explanation why the Simpson Paradox occurs.
}
\description{
`compare` shows two scatterplot to explain the Simpson Paradox. The text in the console explains the concept.
}
\examples{
explore_group(dash)
}
