% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dash}
\alias{dash}
\title{Smoking behaviour and its impact on sprint speed}
\format{
## `dash`
A data frame with 60 rows and 4 columns:
\describe{
  \item{case_id}{Case ID for each student}
  \item{cigarettes_per_day}{The number of cigarettes smoked per day}
  \item{time_100_meter}{The time needed to complete the 100-meter dash in seconds}
  \item{sex}{The sex of each student}
  ...
}
}
\source{
The data is fictitious. It was made up for demonstration reasons.
}
\usage{
dash
}
\description{
An example dataset that contains information on students and their smoking behaviour, their sex and the time they need for the 100-meter dash.
}
\keyword{datasets}
