% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{explore}
\alias{explore}
\title{Explore the datasets}
\usage{
explore(df)
}
\arguments{
\item{df}{(dataframe dash or wages)}
}
\value{
A scatterplot and a description text on the console
}
\description{
`explore` generates a scatterplot with a regression line to give you a first impression of the data.
}
\examples{
explore(dash)
}
