% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{explore_group}
\alias{explore_group}
\title{Take a look at the grouped data}
\usage{
explore_group(df)
}
\arguments{
\item{df}{(dataframe dash or wages)}
}
\value{
A scatterplot and an interpretation of the results.
}
\description{
`explore_group` generates a scatterplot with several regression lines. Color is used to distinguish different groups from each other.
}
\examples{
explore_group(dash)
}
