% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hello_simpson}
\alias{hello_simpson}
\title{Get started with Simpson Paradox}
\usage{
hello_simpson()
}
\value{
A gentle introduction to the simpson package
}
\description{
`hello_simpson` helps you to get started with the simpson-package. You don't need any input.
}
\examples{
hello_simpson()

}
