% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{regression_all}
\alias{regression_all}
\title{Confirm the results from the scatterplot}
\usage{
regression_all(df)
}
\arguments{
\item{df}{(dataframe dash or wages)}
}
\value{
A summary of the regression analysis and an interpretation of the results.
}
\description{
`regression_all` performs a regression analysis on the example data. All data will be considered in the analysis.
}
\examples{
regression_all(dash)
}
