% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{simpson-package}
\alias{simpson}
\alias{simpson-package}
\title{simpson: The Simpson Paradox}
\description{
With the help of this package you can learn more about the Simpson Paradox. This package provides sample data and some functions that help you explore this statistical phenomenon. Start your learning journey with the function hello_simpson().
}
\keyword{internal}
