% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages}
\alias{wages}
\title{The development of wages in the US}
\format{
## `wages`
A data frame with 140 rows and 4 columns:
\describe{
  \item{case_id}{Case ID for each citizen}
  \item{wage_per_month}{Income per month}
  \item{year}{Year}
  \item{education}{Educational profile}
  ...
}
}
\source{
The data is fictitious. It was made up for demonstration reasons.
}
\usage{
wages
}
\description{
An example dataset that contains data on the development of US wages from 2005 to 2014 in different educational groups.
}
\keyword{datasets}
